[back_assignments] move CollectionOfProxy class in its own file
