+++ lib/couchrest_model.rb
@@ -40,6 +40,7 @@ require "couchrest/model/document_queries"
 require "couchrest/model/extended_attachments"
 require "couchrest/model/proxyable"
 require "couchrest/model/associations"
+require "couchrest/model/associations/collection_of_proxy"
 require "couchrest/model/configuration"
 require "couchrest/model/connection"
 require "couchrest/model/design"
